test(p2-shim): group browser tests by interface
