+++ packages/preview2-shim/test/browser/cli.ts
@@ -1,0 +1,28 @@
+import { suite, test, assert } from "vitest";
+
+suite("Browser CLI", () => {
+    test("browser CLI factories isolate configuration and streams", async () => {
+        const { createCli } = await import("../../src/browser/cli.js");
+        const firstWrites: number[] = [];
+        const secondWrites: number[] = [];
+        const first = createCli({
+            environment: { INSTANCE: "first" },
+            arguments: ["one"],
+            stdout: { write: (bytes) => firstWrites.push(...bytes) },
+        });
+        const second = createCli({
+            environment: { INSTANCE: "second" },
+            arguments: ["two"],
+            stdout: { write: (bytes) => secondWrites.push(...bytes) },
+        });
+
+        assert.deepStrictEqual(first.environment.getEnvironment(), [["INSTANCE", "first"]]);
+        assert.deepStrictEqual(second.environment.getEnvironment(), [["INSTANCE", "second"]]);
+        const firstStdout = first.stdout.getStdout();
+        firstStdout.checkWrite();
+        firstStdout.write(new Uint8Array([1, 2]));
+        assert.deepStrictEqual(firstWrites, [1, 2]);
+        assert.deepStrictEqual(secondWrites, []);
+        assert.strictEqual(first.terminalStdout.getTerminalStdout(), undefined);
+    });
+});
